+++ src/ui/mixins/tests/empty-list-test.js
@@ -53,7 +53,8 @@ describe("Empty List Mixin", function() {
 
   afterEach(function() {
     document.body.removeChild(testRoot);
-    if (el) el.onDestroy();
+    if (el) el.destroy();
+    if (client) client.destroy();
     jasmine.clock().uninstall();
 
   });