+++ src/ui/mixins/tests/list-item-selection-test.js
@@ -28,6 +28,7 @@ describe("List Item Selection Mixin", function() {
     jasmine.clock().uninstall();
     document.body.removeChild(testRoot);
 
+    if (el) el.destroy();
     if (client) client.destroy();
   });