+++ test/core/integration/conversationIntegrationSpec.js
@@ -58,10 +58,17 @@ describe("Conversation Integration Tests", function() {
         });
 
         jasmine.clock().tick(1);
+        Layer.Utils.defer.flush();
         requests.reset();
         syncManager.queue = [request];
         client.syncManager.queue = [];
         client._clientReady();
+    });
+
+    afterEach(function() {
+      jasmine.clock().uninstall();
+      jasmine.Ajax.uninstall();
+      if (client) client.destroy();
     });
 
     afterAll(function() {